+++ .atomist/editors/AddLocalEditor.ts
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-import { EditProject } from '@atomist/rug/operations/ProjectEditor';
-import { Project } from '@atomist/rug/model/Project';
-import { Pattern } from '@atomist/rug/operations/RugOperation';
-import { Editor, Parameter, Tags } from '@atomist/rug/operations/Decorators';
+import { Project } from "@atomist/rug/model/Project";
+import { Editor, Parameter, Tags } from "@atomist/rug/operations/Decorators";
+import { EditProject } from "@atomist/rug/operations/ProjectEditor";
+import { Pattern } from "@atomist/rug/operations/RugOperation";
 
-import { RugParameters } from './RugParameters';
+import { RugParameters } from "./RugParameters";
 
 @Editor("AddLocalEditor", "adds an editor for modifying the local project, initiating a Rug archive if needed")
 @Tags("rug", "atomist", "typescript")
@@ -28,28 +28,32 @@ export class AddLocalEditor implements EditProject {
     @Parameter({
         ...RugParameters.Name,
         name: "Editor Name",
-        description: "name for the new Rug editor"
+        description: "name for the new Rug editor",
     })
-    editorName: string;
+    public editorName: string;
 
     @Parameter({
         ...RugParameters.Description,
         name: "Description",
         description: "description of the new Rug editor",
-        required: false
+        required: false,
     })
-    description: string = "an editor for modifying this project";
+    public description: string = "an editor for modifying this project";
 
     @Parameter(RugParameters.GroupId)
-    groupId: string = "local";
+    public groupId: string = "local";
 
-    edit(project: Project) {
-        project.editWith("ConvertExistingProjectToRugArchive", { description: this.description, archiveName: project.name, groupId: this.groupId });
+    public edit(project: Project) {
+        project.editWith("ConvertExistingProjectToRugArchive", {
+            description: this.description,
+            archiveName: project.name,
+            groupId: this.groupId,
+        });
         project.editWith("AddTypeScript", {});
         if (!project.directoryExists(".atomist/node_modules")) {
             // The following line works with the CLI but when used through the
             // bot, it often triggers GitHub rate limiting.
-            //project.copyEditorBackingFilesPreservingPath(".atomist/node_modules");
+            // project.copyEditorBackingFilesPreservingPath(".atomist/node_modules");
         }
 
         project.editWith("AddTypeScriptEditor", this);