+++ .atomist/editors/HavePathExpressionEngine.ts
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-import { EditProject } from '@atomist/rug/operations/ProjectEditor';
-import { Project } from '@atomist/rug/model/Project';
-import { PathExpression, PathExpressionEngine } from '@atomist/rug/tree/PathExpression';
-import { Pattern } from '@atomist/rug/operations/RugOperation';
-import { File } from '@atomist/rug/model/File';
-import { Editor, Parameter, Tags } from '@atomist/rug/operations/Decorators';
+import { File } from "@atomist/rug/model/File";
+import { Project } from "@atomist/rug/model/Project";
+import { Editor, Parameter, Tags } from "@atomist/rug/operations/Decorators";
+import { EditProject } from "@atomist/rug/operations/ProjectEditor";
+import { Pattern } from "@atomist/rug/operations/RugOperation";
+import { PathExpression, PathExpressionEngine } from "@atomist/rug/tree/PathExpression";
 
 @Editor("HavePathExpressionEngine", "adds access to path expression engine in a Rug editor")
 @Tags("atomist", "rug")
@@ -31,25 +31,27 @@ export class HavePathExpressionEngine implements EditProject {
         pattern: Pattern.any,
         validInput: "an existing TypeScript rug in your repository",
         minLength: 1,
-        maxLength: 100
+        maxLength: 100,
     })
-    rugName: string;
+    public rugName: string;
 
-    edit(project: Project) {
-        let rugPath = `.atomist/editors/${this.rugName}.ts`;
-        let rug: File = project.findFile(rugPath);
+    public edit(project: Project) {
+        const rugPath = `.atomist/editors/${this.rugName}.ts`;
+        const rug: File = project.findFile(rugPath);
         if (rug == null) {
             console.log(`did not find ${rugPath}`);
             return;
         }
-        console.log(`Found file ${rug.path}`)
+        console.log(`Found file ${rug.path}`);
         if (!rug.contains("project.context().pathExpressionEngine()")) {
             // microgrammars would look better. Need rug 0.13
-            rug.regexpReplace("(edit\\(.*\\) \\{)", "$1\n        let eng: PathExpressionEngine = project.context().pathExpressionEngine();\n");
+            rug.regexpReplace("(edit\\(.*\\) \\{)",
+                "$1\n        const eng: PathExpressionEngine = project.context.pathExpressionEngine;\n");
         }
-        if (!rug.contains("import { PathExpression, PathExpressionEngine } from '@atomist/rug/tree/PathExpression'")) {
+        const pxeImport = `import { PathExpression, PathExpressionEngine } from "@atomist/rug/tree/PathExpression"`;
+        if (!rug.contains(pxeImport)) {
             // if we had a TypeScript language extension, "add import" would be a thing
-            rug.prepend("import { PathExpression, PathExpressionEngine } from '@atomist/rug/tree/PathExpression'\n");
+            rug.prepend(pxeImport + "\n");
         }
     }
 }